refactor: move 3D view lifecycle into volume viewer
